Refactored CacheService, changed package.
