+++ src/main/java/org/java2uml/java2umlapi/fileStorage/service/ClassDiagramSVGService.java
@@ -1,5 +1,6 @@
 package org.java2uml.java2umlapi.fileStorage.service;
 
+import org.java2uml.java2umlapi.util.cache.CacheService;
 import org.springframework.stereotype.Service;
 
 /**
